+++ tests/components/ports/test_mw_fem_analog_ports.py
@@ -1,4 +1,5 @@
 import pytest
+import json
 
 from quam.components.ports.analog_outputs import MWFEMAnalogOutputPort
 from quam.components.ports.analog_inputs import MWFEMAnalogInputPort
@@ -63,7 +64,7 @@ def test_mw_fem_analog_output_port():
                                 "sampling_rate": 1e9,
                                 "full_scale_power_dbm": -11,
                             }
-                        }
+                        },
                     }
                 }
             }
@@ -121,9 +122,34 @@ def test_mw_fem_analog_input_ports():
                                 "sampling_rate": 1e9,
                                 "shareable": False,
                             }
-                        }
+                        },
                     }
                 }
             }
         }
     }
+
+
+def test_mw_fem_upconverter_config_generation():
+    port = MWFEMAnalogOutputPort(
+        "con1",
+        1,
+        2,
+        band=1,
+        upconverters={1: {"frequency": 4.5e9}, 2: {"frequency": 5.5e9}},
+    )
+    cfg = {"controllers": {}}
+    port.apply_to_config(cfg)
+
+    assert cfg["controllers"]["con1"]["fems"][1]["analog_outputs"][2] == {
+        "band": 1,
+        "delay": 0,
+        "shareable": False,
+        "sampling_rate": 1000000000.0,
+        "full_scale_power_dbm": -11,
+        "upconverters": {
+            1: {"frequency": 4500000000.0},
+            2: {"frequency": 5500000000.0},
+        },
+    }
+    json.dumps(cfg, indent=4)